% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_LW.R
\name{beta_LW}
\alias{beta_LW}
\title{Caculate the estimators of beta on the LEV-opt#'}
\usage{
beta_LW(X, Y, K, nk)
}
\arguments{
\item{X}{is the observation matrix}

\item{Y}{is the response vector}

\item{K}{is the number of subsets}

\item{nk}{is the length of subsets}
}
\value{
A list containing:
\item{betalev}{The estimator of beta on the LEV-opt subset.}
\item{betam}{The mean of the beta estimators across all K subsets.}
\item{AMSE}{The Average Mean Squared Error (AMSE) for the estimator.}
\item{WMSE}{The Weighted Mean Squared Error (WMSE) for the estimator.}
\item{MSElevb}{The Mean Squared Error (MSE) of the LEV-opt estimator compared to the true beta.}
\item{MSEb}{The Mean Squared Error (MSE) of the mean estimator (betam) compared to the true beta.}
\item{MSEyleva}{The Mean Squared Error (MSE) of the LEV-opt estimator on the subset with the maximum hat value (Xleva).}
\item{MSEyleviy}{The Mean Squared Error (MSE) of the LEV-opt estimator on the subset with the minimum hat value (Xlevi).}
\item{MSEW}{The Mean Squared Error (MSE) of the weighted estimator (Wbeta) compared to the true beta.}
\item{MSEw}{The Mean Squared Error (MSE) of the weighted estimator (wbeta) compared to the true beta.}
}
\description{
Caculate the estimators of beta on the LEV-opt#'
}
