\name{ethylene_CO}
\alias{ethylene_CO}
\docType{data}
\title{
The chemical sensor data set
}
\description{
A data set about chemical sensor 
}
\usage{data("ethylene_CO")}
\format{
  A data frame with 4001 observations on the following 19 variables.
  \describe{
    \item{\code{V1}}{a character vector}
    \item{\code{V2}}{a character vector}
    \item{\code{V3}}{a character vector}
    \item{\code{V4}}{a character vector}
    \item{\code{V5}}{a character vector}
    \item{\code{V6}}{a character vector}
    \item{\code{V7}}{a character vector}
    \item{\code{V8}}{a character vector}
    \item{\code{V9}}{a character vector}
    \item{\code{V10}}{a character vector}
    \item{\code{V11}}{a character vector}
    \item{\code{V12}}{a character vector}
    \item{\code{V13}}{a character vector}
    \item{\code{V14}}{a character vector}
    \item{\code{V15}}{a character vector}
    \item{\code{V16}}{a character vector}
    \item{\code{V17}}{a character vector}
    \item{\code{V18}}{a character vector}
    \item{\code{V19}}{a character vector}
  }
}
\details{
We selected the first 4001 rows on the original data set about 1048576 observations on 19 variables.
}
\source{
UCI Repository
}
\references{
Wang, H. Y., Zhu, R., and Ma, P. (2018). Optimal subsampling for large sample logistic regression.
Journal of the American Statistical Association, 113(522), 829-844.
}
\examples{
data(ethylene_CO)
## maybe str(ethylene_CO) ; plot(ethylene_CO) ...
}
\keyword{datasets}
